Fixed several warnings raised by the LLVM compiler.
git-svn-id: https://s.snth.net/svn/rlk/thumb@2714 dcf3a027-e0e4-0310-89c9-ab9954db6d00

+++ src/ogl-program.cpp
@@ -53,8 +53,8 @@ GLuint ogl::program::compile(GLenum type, const std::string& name,
     {
         handle = glCreateShader(type);
 
-        const char *data = text.data();
-        GLint       size = text.size();
+        const char *data =       text.data();
+        GLint       size = GLint(text.size());
 
         glShaderSource (handle, 1, &data, &size);
         glCompileShader(handle);
